something

+++ src/br/com/alura/minhasmusicas/modelos/Audio.java
@@ -4,7 +4,7 @@ public class Audio {
     private String titulo;
     private int totalReproducoes;
     private int totalCurtidas;
-    private double classificacao;
+    private int classificacao;
 
     public void curte() {
         this.totalCurtidas ++;
@@ -14,24 +14,29 @@ public class Audio {
         this.totalReproducoes ++;
     }
 
-    public double getClassificacao() {
+    public int getClassificacao() {
+
         return classificacao;
     }
 
     public String getTitulo() {
+
         return titulo;
     }
 
     public void setTitulo(String titulo) {
+
         this.titulo = titulo;
     }
 
     public int getTotalReproducoes() {
+
         return totalReproducoes;
     }
 
 
     public int getTotalCurtidas() {
+
         return totalCurtidas;
     }
